requirements.txt
django
django-registration-redux
djangorestframework
ipython
raven
